Delete old files that were reworked and rename the reworks to replace them.
