+++ src/JHG_inspector/presentation_layer/Container.py
@@ -1,130 +1,101 @@
-from functools import partial
+from typing import Union, Optional
 
 from PyQt6.QtCore import Qt
-from PyQt6.QtGui import QContextMenuEvent
-from PyQt6.QtWidgets import QSplitter, QWidget, QMenu, QLabel
+from PyQt6.QtWidgets import QSplitter
 
-from src.JHG_inspector.presentation_layer.components.TabbedPanels import TabbedPanels
+from src.JHG_inspector.presentation_layer.PanelTabWidget import PanelTabWidget
+from src.JHG_inspector.presentation_layer.panels.Panel import Panel
 
 
 class Container(QSplitter):
-    """Used to create a runtime customizable layout.
+    """Used to create custom layouts on the fly.
 
-       Each Container can hold up to two items (which can be either TabbedPanel objects, other Containers, or one
-       of each). They are either displayed vertically or horizontally, based on the orientation passed at construction.
-       By calling the split command, you can add either the second item, or if two already exist, replace an item with
-       a new container with its first item being the replaced item. If both items are removed, the Container is deleted.
+       Each Container can hold a maximum of two items (which is either another Container object or a PanelTabWidget).
+       If another item is added, a new container is made and put in that place.
        """
-
-    num_containers = 0
-
-    def __init__(self, orientation: Qt.Orientation = Qt.Orientation.Horizontal):
-        """
-        Parameters
-        ----------
-        orientation: Qt.Orientation
-            Defines whether the items will be displayed one on top of the other (Qt.Orientation.Vertical), or side
-            by side (Qt.Orientation.Horizontal).
-        """
-
-        Container.num_containers += 1
+    num_containers = 1
+    def __init__(self, item: Union["Container", "PanelTabWidget"], parent_container: "Container" = None,
+                 split_direction: Qt.Orientation = Qt.Orientation.Horizontal):
+        super().__init__(parent_container)
         self.container_num = Container.num_containers
-        super().__init__()
-        self.setOrientation(orientation)
-        self.setHandleWidth(1)
+        Container.num_containers += 1
 
-    def add_child(self, widget: QWidget, split_direction: Qt.Orientation = Qt.Orientation.Horizontal, index: int = 1):
-        """Adds a widget to a Container, nesting a new container if necessary.
+        self.parent_container = parent_container
+        self.items = [item, None]
+        self.setOrientation(split_direction)
 
-           Adds a widget if there is only one item in the Container. Otherwise, creates a new Container, places the item
-           at the index to be split in the new Container, adds the passed widget to the new Container, a replaces the
-           item at the index to be split with the new Container.
+        self.addWidget(item)
+        self.setCollapsible(0, False)
 
-           Parameters
-           ----------
-           widget: QWidget
-               The new widget to be added
-           index: int
-               The index of the parent Container where the new widget should be placed (usually through instantiating a
-               new Container and placing the widget in it)
-           split_direction
-               The orientation (horizontal or vertical) that the container the widget will be placed in should display
-               its items. If the parent Container has only one item, that is the affected container. If it already has
-               two, then the split_direction is applied to the new Container.
+    def add_item(self, new_item: Union["Container", "PanelTabWidget"],
+                 split_direction: Qt.Orientation = Qt.Orientation.Horizontal, existing_item: Optional[Union["Container", "PanelTabWidget"]] = None):
+        """Adds a new item to the container.
+
+           If there are already two items in the Container, a new Container is created, the panel that was already there
+           is added to the new Container, and that panel is replaced with the new Container.
            """
 
-        if self.count() == 1:
+        if not self.items[1]:
+            self.items[1] = new_item
+            self.addWidget(new_item)
             self.setOrientation(split_direction)
-            self.addWidget(widget)
             self.setCollapsible(1, False)
-            self.setStretchFactor(self.indexOf(widget), 1)
+        elif existing_item:
+            new_container = Container(existing_item, self, split_direction)
+            self.replace_item(existing_item, new_container)
+            new_container.add_item(new_item, split_direction)
+            existing_item.parent_container = new_container
+            new_item.parent_container = new_container
 
-            if isinstance(widget, TabbedPanels):
-                widget.parent_container = self
         else:
-            old_widget = self.widget(index)  # The widget at the position where the split is to occur
+            raise ValueError("Panel must be provided when both items are occupied — the Container needs to know which "
+                             "side to split.")
 
+    def remove_item(self, item: Union["Container", "PanelTabWidget"]):
+        """Removes the item from the container, and removes itself if it is now empty."""
+        if item not in self.items:
+            raise ValueError("Item not in this Container.")
 
-            nested_container = Container(orientation=split_direction)
-            if old_widget:
-                old_widget.setParent(nested_container)
-                nested_container.addWidget(old_widget)
-                nested_container.addWidget(widget)
-                nested_container.setCollapsible(1, False)
+        item.setParent(None)
+        item.deleteLater()
+
+        if item == self.items[0]:
+            # If the Container has two elements, promote the 2nd item to the first item. If not, the Container will be
+            # empty following the deletion of the 1st item, and should be removed from the parent Container
+            if self.items[1]:
+                promoted = self.items[1]
+                self.items[0] = promoted
+                self.items[1] = None
+            elif self.parent_container: # Makes it so that a CentralContainer doesn't call parent_container.remove_item
+                self.parent_container.remove_item(self)
             else:
-                nested_container.addWidget(widget)
+                self.items[0] = None
+        else:
+            self.items[1] = None
 
+    def replace_item(self, old_item: Union["Container", "Panel"], new_item: Union["Container", "Panel"]):
+        """Replaces an item with a new one. Used for both collapsing structures of Containers that only hold other
+           Containers and for putting new containers in the place of panels that will be split."""
+        index = self.items.index(old_item)
+        if index == -1:
+            raise ValueError("Item not found in this Container.")
 
-            nested_container.setCollapsible(0, False)
+        old_item.setParent(new_item)
+        new_item.addWidget(old_item)
 
-            if isinstance(widget, TabbedPanels):
-                widget.parent_container = nested_container
+        self.addWidget(new_item)
+        self.items[index] = new_item
 
-            self.insertWidget(index, nested_container)
-            self.setStretchFactor(self.indexOf(widget), 1)
-
-    def empty_check(self):
-        """A call back called by children Containers and TabbedPanels when they are cleared.
-
-           Allows for Container's items to let the Container know when it should check if it is empty. If it is, it then
-           propagates the empty_check up to its parent, allowing for all ancestor Containers to close once they have
-           no more items in their tree.
-           """
-
-        if self.count() == 0:
-            Container.num_containers -= 1
-
-            parent_container = self.find_parent_container()
-
-            self.setParent(None)
-            self.deleteLater()
-
-            if parent_container is not None:
-                parent_container.empty_check()
-
-    def find_parent_container(self):
-        parent = self.parent()
-        while parent and not isinstance(parent, Container):
-            parent = parent.parent()
-        return parent
-
-    def contextMenuEvent(self, event: QContextMenuEvent):
-        widget_under_mouse = self.childAt(event.pos())
-        while widget_under_mouse and widget_under_mouse.parent() != self:
-            widget_under_mouse = widget_under_mouse.parent()
-        index = self.indexOf(widget_under_mouse)
-
-        menu = QMenu(self)
-        split_right = menu.addAction("Split Right")
-        split_right.triggered.connect(partial(self.split, split_direction = Qt.Orientation.Horizontal, splitter_index=index))
-
-        split_down = menu.addAction("Split Down")
-        split_down.triggered.connect(partial(self.split, split_direction = Qt.Orientation.Vertical, splitter_index=index))
-
-        menu.exec(event.globalPos())
-
-    def split(self, widget: QWidget = None, split_direction: Qt.Orientation = Qt.Orientation.Horizontal, splitter_index=2):
-        if not widget:
-            widget = QLabel("New Widget")
-        tabs = TabbedPanels(self.empty_check, new_widget = widget)
-        self.add_child(tabs, split_direction, splitter_index)
+    def focus_first_panel(self):
+        if isinstance(self.items[0], PanelTabWidget):
+            self.items[0].widget(0).setFocus()
+            return True
+        elif isinstance(self.items[1], PanelTabWidget):
+            self.items[1].widget(0).setFocus()
+            return True
+        elif self.items[0].focus_first_panel():
+            return True
+        elif self.items[1].focus_first_panel():
+            return True
+        else:
+            return False
