+++ src/JHG_inspector/presentation_layer/MainWindow.py
@@ -6,10 +6,10 @@ from PyQt6.QtWidgets import QMainWindow, QFileDialog, QStatusBar, QSplitter
 
 from src.JHG_inspector.logic_layer.DatabaseManager import DatabaseManager
 from src.JHG_inspector.logic_layer.ToolsManager import ToolsManager
-from src.JHG_inspector.presentation_layer.ContainerRework import Container
+from src.JHG_inspector.presentation_layer.Container import Container
 from src.JHG_inspector.presentation_layer.dialogs.GamesDialog import GamesDialog
 from src.JHG_inspector.presentation_layer.dialogs.OpenToolDialog import OpenToolDialog
-from src.JHG_inspector.presentation_layer.panels.CentralContainerRework import CentralContainer
+from src.JHG_inspector.presentation_layer.panels.CentralContainer import CentralContainer
 from src.JHG_inspector.presentation_layer.panels.GamesetPanel import GamesetPanel
 from src.JHG_inspector.presentation_layer.panels.Panel import Panel
